Try to rename field folder
