+++ packages/ui/src/SelectField.tsx
@@ -2,7 +2,7 @@ import React, {FC} from "react";
 import {View} from "react-native";
 
 import {SelectFieldProps} from "./Common";
-import {FieldError, FieldHelperText, FieldTitle} from "./fieldElements";
+import {FieldError, FieldHelperText, FieldTitle} from "./fieldElems";
 import {RNPickerSelect} from "./PickerSelect";
 
 export const SelectField: FC<SelectFieldProps> = ({
